[refactor] rename ent->entities
