+++ adapters/repos/db/lsmkv/bucket.go
@@ -24,7 +24,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/sirupsen/logrus"
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/entities/cyclemanager"
 	"github.com/weaviate/weaviate/entities/storagestate"
 	"github.com/weaviate/weaviate/entities/storobj"
@@ -171,13 +171,13 @@ func (b *Bucket) Get(key []byte) ([]byte, error) {
 		// is replace
 		return v, nil
 	}
-	if err == ent.Deleted {
+	if err == entities.Deleted {
 		// deleted in the mem-table (which is always the latest) means we don't
 		// have to check the disk segments, return nil now
 		return nil, nil
 	}
 
-	if err != ent.NotFound {
+	if err != entities.NotFound {
 		panic("unsupported error in bucket.Get")
 	}
 
@@ -188,13 +188,13 @@ func (b *Bucket) Get(key []byte) ([]byte, error) {
 			// is replace
 			return v, nil
 		}
-		if err == ent.Deleted {
+		if err == entities.Deleted {
 			// deleted in the now most recent memtable  means we don't have to check
 			// the disk segments, return nil now
 			return nil, nil
 		}
 
-		if err != ent.NotFound {
+		if err != entities.NotFound {
 			panic("unsupported error in bucket.Get")
 		}
 	}
@@ -223,13 +223,13 @@ func (b *Bucket) GetBySecondary(pos int, key []byte) ([]byte, error) {
 		// is replace
 		return v, nil
 	}
-	if err == ent.Deleted {
+	if err == entities.Deleted {
 		// deleted in the mem-table (which is always the latest) means we don't
 		// have to check the disk segments, return nil now
 		return nil, nil
 	}
 
-	if err != ent.NotFound {
+	if err != entities.NotFound {
 		panic("unsupported error in bucket.Get")
 	}
 
@@ -240,13 +240,13 @@ func (b *Bucket) GetBySecondary(pos int, key []byte) ([]byte, error) {
 			// is replace
 			return v, nil
 		}
-		if err == ent.Deleted {
+		if err == entities.Deleted {
 			// deleted in the now most recent memtable  means we don't have to check
 			// the disk segments, return nil now
 			return nil, nil
 		}
 
-		if err != ent.NotFound {
+		if err != entities.NotFound {
 			panic("unsupported error in bucket.Get")
 		}
 	}
@@ -266,7 +266,7 @@ func (b *Bucket) SetList(key []byte) ([][]byte, error) {
 
 	v, err := b.disk.getCollection(key)
 	if err != nil {
-		if err != nil && err != ent.NotFound {
+		if err != nil && err != entities.NotFound {
 			return nil, err
 		}
 	}
@@ -275,7 +275,7 @@ func (b *Bucket) SetList(key []byte) ([][]byte, error) {
 	if b.flushing != nil {
 		v, err = b.flushing.getCollection(key)
 		if err != nil {
-			if err != nil && err != ent.NotFound {
+			if err != nil && err != entities.NotFound {
 				return nil, err
 			}
 		}
@@ -285,7 +285,7 @@ func (b *Bucket) SetList(key []byte) ([][]byte, error) {
 
 	v, err = b.active.getCollection(key)
 	if err != nil {
-		if err != nil && err != ent.NotFound {
+		if err != nil && err != entities.NotFound {
 			return nil, err
 		}
 	}
@@ -408,7 +408,7 @@ func (b *Bucket) MapList(key []byte, cfgs ...MapListOption) ([]MapPair, error) {
 	// before := time.Now()
 	disk, err := b.disk.getCollectionBySegments(key)
 	if err != nil {
-		if err != nil && err != ent.NotFound {
+		if err != nil && err != entities.NotFound {
 			return nil, err
 		}
 	}
@@ -432,7 +432,7 @@ func (b *Bucket) MapList(key []byte, cfgs ...MapListOption) ([]MapPair, error) {
 	if b.flushing != nil {
 		v, err := b.flushing.getMap(key)
 		if err != nil {
-			if err != nil && err != ent.NotFound {
+			if err != nil && err != entities.NotFound {
 				return nil, err
 			}
 		}
@@ -443,7 +443,7 @@ func (b *Bucket) MapList(key []byte, cfgs ...MapListOption) ([]MapPair, error) {
 	// before = time.Now()
 	v, err := b.active.getMap(key)
 	if err != nil {
-		if err != nil && err != ent.NotFound {
+		if err != nil && err != entities.NotFound {
 			return nil, err
 		}
 	}
