+++ adapters/repos/db/lsmkv/cursor_bucket_set.go
@@ -15,7 +15,7 @@ import (
 	"bytes"
 
 	"github.com/pkg/errors"
-	"github.com/semi-technologies/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/semi-technologies/weaviate/adapters/repos/db/lsmkv/entities"
 )
 
 type CursorSet struct {
@@ -102,7 +102,7 @@ func (c *CursorSet) seekAll(target []byte) {
 	state := make([]cursorStateCollection, len(c.innerCursors))
 	for i, cur := range c.innerCursors {
 		key, value, err := cur.seek(target)
-		if err == ent.NotFound {
+		if err == entities.NotFound {
 			state[i].err = err
 			continue
 		}
@@ -124,7 +124,7 @@ func (c *CursorSet) firstAll() {
 	state := make([]cursorStateCollection, len(c.innerCursors))
 	for i, cur := range c.innerCursors {
 		key, value, err := cur.first()
-		if err == ent.NotFound {
+		if err == entities.NotFound {
 			state[i].err = err
 			continue
 		}
@@ -145,13 +145,13 @@ func (c *CursorSet) firstAll() {
 func (c *CursorSet) serveCurrentStateAndAdvance() ([]byte, [][]byte) {
 	id, err := c.cursorWithLowestKey()
 	if err != nil {
-		if err == ent.NotFound {
+		if err == entities.NotFound {
 			return nil, nil
 		}
 	}
 
 	// check if this is a duplicate key before checking for the remaining errors,
-	// as cases such as 'ent.Deleted' can be better handled inside
+	// as cases such as 'entities.Deleted' can be better handled inside
 	// mergeDuplicatesInCurrentStateAndAdvance where we can be sure to act on
 	// segments in the correct order
 	if ids, ok := c.haveDuplicatesInState(id); ok {
@@ -162,12 +162,12 @@ func (c *CursorSet) serveCurrentStateAndAdvance() ([]byte, [][]byte) {
 }
 
 func (c *CursorSet) cursorWithLowestKey() (int, error) {
-	err := ent.NotFound
+	err := entities.NotFound
 	pos := -1
 	var lowest []byte
 
 	for i, res := range c.state {
-		if res.err == ent.NotFound {
+		if res.err == entities.NotFound {
 			continue
 		}
 
@@ -227,7 +227,7 @@ func (c *CursorSet) mergeDuplicatesInCurrentStateAndAdvance(ids []int) ([]byte, 
 
 func (c *CursorSet) advanceInner(id int) {
 	k, v, err := c.innerCursors[id].next()
-	if err == ent.NotFound {
+	if err == entities.NotFound {
 		c.state[id].err = err
 		c.state[id].key = nil
 		if !c.keyOnly {
@@ -236,7 +236,7 @@ func (c *CursorSet) advanceInner(id int) {
 		return
 	}
 
-	if err == ent.Deleted {
+	if err == entities.Deleted {
 		c.state[id].err = err
 		c.state[id].key = k
 		c.state[id].value = nil
