+++ adapters/repos/db/lsmkv/cursor_segment_collection.go
@@ -12,7 +12,7 @@
 package lsmkv
 
 import (
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/segmentindex"
 )
 
@@ -42,7 +42,7 @@ func (s *segmentCursorCollection) seek(key []byte) ([]byte, []value, error) {
 	node, err := s.segment.index.Seek(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
-			return nil, nil, ent.NotFound
+			return nil, nil, entities.NotFound
 		}
 
 		return nil, nil, err
@@ -52,7 +52,7 @@ func (s *segmentCursorCollection) seek(key []byte) ([]byte, []value, error) {
 		s.segment.contents[node.Start:node.End])
 
 	// make sure to set the next offset before checking the error. The error
-	// could be 'ent.Deleted' which would require that the offset is still advanced
+	// could be 'entities.Deleted' which would require that the offset is still advanced
 	// for the next cycle
 	s.nextOffset = node.End
 	if err != nil {
@@ -64,14 +64,14 @@ func (s *segmentCursorCollection) seek(key []byte) ([]byte, []value, error) {
 
 func (s *segmentCursorCollection) next() ([]byte, []value, error) {
 	if s.nextOffset >= s.segment.dataEndPos {
-		return nil, nil, ent.NotFound
+		return nil, nil, entities.NotFound
 	}
 
 	parsed, err := s.segment.collectionStratParseDataWithKey(
 		s.segment.contents[s.nextOffset:])
 
 	// make sure to set the next offset before checking the error. The error
-	// could be 'ent.Deleted' which would require that the offset is still advanced
+	// could be 'entities.Deleted' which would require that the offset is still advanced
 	// for the next cycle
 	s.nextOffset = s.nextOffset + uint64(parsed.offset)
 	if err != nil {
@@ -87,7 +87,7 @@ func (s *segmentCursorCollection) first() ([]byte, []value, error) {
 		s.segment.contents[s.nextOffset:])
 
 	// make sure to set the next offset before checking the error. The error
-	// could be 'ent.Deleted' which would require that the offset is still advanced
+	// could be 'entities.Deleted' which would require that the offset is still advanced
 	// for the next cycle
 	s.nextOffset = s.nextOffset + uint64(parsed.offset)
 	if err != nil {
