+++ adapters/repos/db/lsmkv/cursor_segment_map.go
@@ -12,7 +12,7 @@
 package lsmkv
 
 import (
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/segmentindex"
 )
 
@@ -42,7 +42,7 @@ func (s *segmentCursorMap) seek(key []byte) ([]byte, []MapPair, error) {
 	node, err := s.segment.index.Seek(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
-			return nil, nil, ent.NotFound
+			return nil, nil, entities.NotFound
 		}
 
 		return nil, nil, err
@@ -72,7 +72,7 @@ func (s *segmentCursorMap) seek(key []byte) ([]byte, []MapPair, error) {
 
 func (s *segmentCursorMap) next() ([]byte, []MapPair, error) {
 	if s.nextOffset >= s.segment.dataEndPos {
-		return nil, nil, ent.NotFound
+		return nil, nil, entities.NotFound
 	}
 
 	parsed, err := s.segment.collectionStratParseDataWithKey(
