+++ adapters/repos/db/lsmkv/cursor_segment_replace.go
@@ -12,7 +12,7 @@
 package lsmkv
 
 import (
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/segmentindex"
 )
 
@@ -44,7 +44,7 @@ func (s *segmentCursorReplace) seek(key []byte) ([]byte, []byte, error) {
 	node, err := s.segment.index.Seek(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
-			return nil, nil, ent.NotFound
+			return nil, nil, entities.NotFound
 		}
 
 		return nil, nil, err
@@ -67,7 +67,7 @@ func (s *segmentCursorReplace) seek(key []byte) ([]byte, []byte, error) {
 
 func (s *segmentCursorReplace) next() ([]byte, []byte, error) {
 	if s.nextOffset >= s.segment.dataEndPos {
-		return nil, nil, ent.NotFound
+		return nil, nil, entities.NotFound
 	}
 
 	err := s.segment.replaceStratParseDataWithKeyInto(
@@ -103,7 +103,7 @@ func (s *segmentCursorReplace) first() ([]byte, []byte, error) {
 func (s *segmentCursorReplace) nextWithAllKeys() (segmentReplaceNode, error) {
 	out := segmentReplaceNode{}
 	if s.nextOffset >= s.segment.dataEndPos {
-		return out, ent.NotFound
+		return out, entities.NotFound
 	}
 
 	parsed, err := s.segment.replaceStratParseDataWithKey(
