+++ adapters/repos/db/lsmkv/segment_collection_strategy.go
@@ -16,7 +16,7 @@ import (
 	"encoding/binary"
 
 	"github.com/pkg/errors"
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/segmentindex"
 )
 
@@ -28,13 +28,13 @@ func (i *segment) getCollection(key []byte) ([]value, error) {
 	}
 
 	if !i.bloomFilter.Test(key) {
-		return nil, ent.NotFound
+		return nil, entities.NotFound
 	}
 
 	node, err := i.index.Get(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
-			return nil, ent.NotFound
+			return nil, entities.NotFound
 		} else {
 			return nil, err
 		}
@@ -61,7 +61,7 @@ func (i *segment) getCollection(key []byte) ([]value, error) {
 
 func (i *segment) collectionStratParseData(in []byte) ([]value, error) {
 	if len(in) == 0 {
-		return nil, ent.NotFound
+		return nil, entities.NotFound
 	}
 
 	offset := 0
@@ -91,7 +91,7 @@ func (i *segment) collectionStratParseDataWithKey(in []byte) (segmentCollectionN
 	r := bytes.NewReader(in)
 
 	if len(in) == 0 {
-		return segmentCollectionNode{}, ent.NotFound
+		return segmentCollectionNode{}, entities.NotFound
 	}
 
 	return ParseCollectionNode(r)
@@ -99,7 +99,7 @@ func (i *segment) collectionStratParseDataWithKey(in []byte) (segmentCollectionN
 
 func (i *segment) collectionStratParseDataWithKeyInto(in []byte, node *segmentCollectionNode) error {
 	if len(in) == 0 {
-		return ent.NotFound
+		return entities.NotFound
 	}
 
 	return ParseCollectionNodeInto(in, node)
