+++ adapters/repos/db/lsmkv/segment_group.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/roaringset"
 	"github.com/weaviate/weaviate/entities/cyclemanager"
 	"github.com/weaviate/weaviate/entities/storagestate"
@@ -189,11 +189,11 @@ func (sg *SegmentGroup) getWithUpperSegmentBoundary(key []byte, topMostSegment i
 	for i := topMostSegment; i >= 0; i-- {
 		v, err := sg.segments[i].get(key)
 		if err != nil {
-			if err == ent.NotFound {
+			if err == entities.NotFound {
 				continue
 			}
 
-			if err == ent.Deleted {
+			if err == entities.Deleted {
 				return nil, nil
 			}
 
@@ -217,11 +217,11 @@ func (sg *SegmentGroup) getBySecondary(pos int, key []byte) ([]byte, error) {
 	for i := len(sg.segments) - 1; i >= 0; i-- {
 		v, err := sg.segments[i].getBySecondary(pos, key)
 		if err != nil {
-			if err == ent.NotFound {
+			if err == entities.NotFound {
 				continue
 			}
 
-			if err == ent.Deleted {
+			if err == entities.Deleted {
 				return nil, nil
 			}
 
@@ -244,7 +244,7 @@ func (sg *SegmentGroup) getCollection(key []byte) ([]value, error) {
 	for _, segment := range sg.segments {
 		v, err := segment.getCollection(key)
 		if err != nil {
-			if err == ent.NotFound {
+			if err == entities.NotFound {
 				continue
 			}
 
@@ -272,7 +272,7 @@ func (sg *SegmentGroup) getCollectionBySegments(key []byte) ([][]value, error) {
 	for _, segment := range sg.segments {
 		v, err := segment.getCollection(key)
 		if err != nil {
-			if err == ent.NotFound {
+			if err == entities.NotFound {
 				continue
 			}
 
@@ -296,7 +296,7 @@ func (sg *SegmentGroup) roaringSetGet(key []byte) (roaringset.BitmapLayers, erro
 	for _, segment := range sg.segments {
 		rs, err := segment.roaringSetGet(key)
 		if err != nil {
-			if err == ent.NotFound {
+			if err == entities.NotFound {
 				continue
 			}
