+++ adapters/repos/db/lsmkv/segment_replace_strategy.go
@@ -17,7 +17,7 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/ent"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/entities"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv/segmentindex"
 )
 
@@ -30,14 +30,14 @@ func (i *segment) get(key []byte) ([]byte, error) {
 
 	if !i.bloomFilter.Test(key) {
 		i.bloomFilterMetrics.trueNegative(before)
-		return nil, ent.NotFound
+		return nil, entities.NotFound
 	}
 
 	node, err := i.index.Get(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
 			i.bloomFilterMetrics.falsePositive(before)
-			return nil, ent.NotFound
+			return nil, entities.NotFound
 		} else {
 			return nil, err
 		}
@@ -72,13 +72,13 @@ func (i *segment) getBySecondary(pos int, key []byte) ([]byte, error) {
 	}
 
 	if !i.secondaryBloomFilters[pos].Test(key) {
-		return nil, ent.NotFound
+		return nil, entities.NotFound
 	}
 
 	node, err := i.secondaryIndices[pos].Get(key)
 	if err != nil {
 		if err == segmentindex.NotFound {
-			return nil, ent.NotFound
+			return nil, entities.NotFound
 		} else {
 			return nil, err
 		}
@@ -103,7 +103,7 @@ func (i *segment) getBySecondary(pos int, key []byte) ([]byte, error) {
 
 func (i *segment) replaceStratParseData(in []byte) ([]byte, error) {
 	if len(in) == 0 {
-		return nil, ent.NotFound
+		return nil, entities.NotFound
 	}
 
 	// byte         meaning
@@ -113,7 +113,7 @@ func (i *segment) replaceStratParseData(in []byte) ([]byte, error) {
 
 	// check the tombstone byte
 	if in[0] == 0x01 {
-		return nil, ent.Deleted
+		return nil, entities.Deleted
 	}
 
 	valueLength := binary.LittleEndian.Uint64(in[1:9])
@@ -123,7 +123,7 @@ func (i *segment) replaceStratParseData(in []byte) ([]byte, error) {
 
 func (i *segment) replaceStratParseDataWithKey(in []byte) (segmentReplaceNode, error) {
 	if len(in) == 0 {
-		return segmentReplaceNode{}, ent.NotFound
+		return segmentReplaceNode{}, entities.NotFound
 	}
 
 	r := bytes.NewReader(in)
@@ -134,7 +134,7 @@ func (i *segment) replaceStratParseDataWithKey(in []byte) (segmentReplaceNode, e
 	}
 
 	if out.tombstone {
-		return out, ent.Deleted
+		return out, entities.Deleted
 	}
 
 	return out, nil
@@ -144,7 +144,7 @@ func (i *segment) replaceStratParseDataWithKeyInto(in []byte,
 	node *segmentReplaceNode,
 ) error {
 	if len(in) == 0 {
-		return ent.NotFound
+		return entities.NotFound
 	}
 
 	r := bytes.NewReader(in)
@@ -155,7 +155,7 @@ func (i *segment) replaceStratParseDataWithKeyInto(in []byte,
 	}
 
 	if node.tombstone {
-		return ent.Deleted
+		return entities.Deleted
 	}
 
 	return nil
